# contracts/openapi/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibraMAS Recommendation API",
    "version": "0.1.0-draft",
    "description": "Versioned contract baseline. Health and explicitly injected G3/G5 slices may be enabled gate by gate; the default runtime remains fail-closed and does not claim that recommendation HTTP is enabled."
  },
  "servers": [
    {"url": "http://localhost:8000", "description": "Local development"}
  ],
...
  "paths": {
...
    "/api/v1/recommendation-impressions/batch": {
      "post": {
        "tags": ["Interaction"],
        "summary": "Append a batch of observed card impressions",
        "operationId": "interaction_append_impressions_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImpressionBatchRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Per-impression idempotent receipts",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImpressionBatchResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
    "/api/v1/recommendation-items/{item_id}/feedback": {
      "post": {
        "tags": ["Interaction"],
        "summary": "Append recommendation feedback",
        "operationId": "interaction_append_feedback_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/ItemId"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FeedbackRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Idempotently replayed feedback",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FeedbackReceipt"}}}
          },
          "201": {
            "description": "Feedback and synchronous profile projection applied",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FeedbackReceipt"}}}
          },
          "202": {
            "description": "Feedback committed and profile projection pending",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FeedbackReceipt"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
    "/api/v1/behavior-events": {
      "post": {
        "tags": ["Interaction"],
        "summary": "Append an allowed independent behavior event",
        "operationId": "interaction_append_behavior_event_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BehaviorEventRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Idempotently replayed event",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BehaviorEventReceipt"}}}
          },
          "201": {
            "description": "Event and synchronous projection applied",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BehaviorEventReceipt"}}}
          },
          "202": {
            "description": "Event committed and profile projection pending",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BehaviorEventReceipt"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"},
      "demoUserAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Demo-User-Id",
        "description": "Local demo identity only. The server accepts it only when APP_ENV=demo and demo identity is explicitly enabled; it never grants research_admin privileges."
      }
    },
    "parameters": {
      "ClientRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {"type": "string", "format": "uuid"},
        "description": "Optional client-generated request UUID. The server validates and echoes it; when absent, the server generates one."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 8, "maxLength": 255},
        "description": "Stable key for safe replay of a write request."
      },
      "DemoUserId": {
        "name": "X-Demo-User-Id",
        "in": "header",
        "required": false,
        "schema": {"type": "integer", "minimum": 1},
        "description": "User-scoped routes only. Accepted only when APP_ENV=demo and demo identity is explicitly enabled; production derives identity from bearer authentication. This header never grants research_admin privileges."
      },
...
      "ItemId": {"name": "item_id", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}},
...
    },
    "headers": {
      "XRequestId": {
        "description": "The validated or server-generated request UUID. It is returned on every operation response.",
        "schema": {"type": "string", "format": "uuid"}
      },
      "XTraceId": {
        "description": "The trace UUID used to correlate the complete operation. It is returned on every operation response.",
        "schema": {"type": "string", "format": "uuid"}
      },
      "IdempotencyReplayed": {
        "description": "True when an idempotent write response was replayed from the original result; false otherwise.",
        "schema": {"type": "boolean"}
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON or invalid HTTP protocol input",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Unauthorized": {
        "description": "Authentication is required or invalid",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Forbidden": {
        "description": "The caller cannot access the requested data or score view",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
...
    },
    "schemas": {
...
      "ImpressionBatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["impressions"],
        "properties": {
          "impressions": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {"$ref": "#/components/schemas/ImpressionInput"}
          }
        }
      },
...
      "ImpressionBatchResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accepted_count", "replayed_count", "rejected_count", "results"],
        "properties": {
          "accepted_count": {"type": "integer", "minimum": 0},
          "replayed_count": {"type": "integer", "minimum": 0},
          "rejected_count": {"type": "integer", "minimum": 0},
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["impression_uuid", "status", "is_valid_exposure"],
              "properties": {
                "impression_uuid": {"type": "string", "format": "uuid"},
                "status": {"type": "string", "enum": ["ACCEPTED", "REPLAYED", "REJECTED"]},
                "is_valid_exposure": {"type": "boolean"},
                "error_code": {"type": ["string", "null"]},
                "agent_action": {"oneOf": [{"$ref": "#/components/schemas/AgentAction"}, {"type": "null"}]}
              }
            }
          }
        }
      },
...
    }
  },
  "security": [{"bearerAuth": []}]
}